+++ tests/test_database.py
@@ -35,8 +35,20 @@ def test_snowflake_connector_password_is_secret_str(connector_params):
 
 def test_snowflake_connector_get_connect_params_get_secret_value(connector_params):
     snowflake_connector = SnowflakeConnector(**connector_params)
-    connector_params = snowflake_connector._get_connect_params()
-    assert connector_params["password"] == "password"
+    connect_params = snowflake_connector._get_connect_params()
+    assert connect_params["password"] == "password"
+
+
+def test_snowflake_connector_get_connect_params_okta_endpoint(connector_params):
+    okta_endpoint = "https://account_name.okta.com"
+    connector_params_okta_endpoint = connector_params.copy()
+    connector_params_okta_endpoint["credentials"].password = None
+    connector_params_okta_endpoint["credentials"].authenticator = "okta_endpoint"
+    connector_params_okta_endpoint["credentials"].okta_endpoint = okta_endpoint
+    snowflake_connector = SnowflakeConnector(**connector_params_okta_endpoint)
+    connect_params = snowflake_connector._get_connect_params()
+    assert connect_params["authenticator"] == okta_endpoint
+    assert connect_params.get("okta_endpoint") is None
 
 
 class SnowflakeCursor: